refactor(web): extract DesignSession, TabStrip, session/pin contexts out of App.tsx (#642 phase 3, seam 5a pt 2)
Verbatim code-move, zero logic edits — only import blocks and `export `
keywords change.

- src/components/session/contexts.ts (new) — ThemeControlContext,
  SessionMeta/SessionsCtx/SessionsContext, PinsCtx/PinsContext, and
  fetchMetrics (a runtime helper both App() and DesignSession call, so it
  needs the same shared home as the contexts rather than living in either
  consumer).
- src/components/session/TabStrip.tsx (new) — TabStrip.
- src/components/session/DesignSession.tsx (new) — WS_URL,
  CONVERGE_N_VALUES, the optimizer-result types (OptMetrics/OptimizeResult/
  OptObjective/OPT_OBJECTIVE_LABELS/OPT_OBJECTIVES/OptPause), and
  DesignSession itself (~3,700 lines). Not split internally — that's seam 5b.
- App.tsx: 4,161 -> 165 lines. What remains is just App() (the shell:
  theme + session-list + pin-list state, its three context providers, and
  the DesignSession mounts) plus its shell-only imports.
